Add controls. Fix a few bugs.
Bugs:
    serial addresses were wrong
    timer missing a break;
    incorrectly being detected as cgb causing cpu_instr to attempt to do a speed switch which stops the cpu.
     incorrectly turning ime off during RST
New:
    implement LCD disable

Tetris now runs, and is "playable". You can see the pieces once they've been placed as they then become part of the BGMap, but you cannot see them when they are falling as they are sprites at that point.

Additionally the game runs way too fast as it's not sync'd to the refresh rate. Up next are window, sprites, and syncing to refresh rate, at which point tetris should be playable.